Moves the tests of elf.rs into a separate file.
